requirements.txt
django-rq
django
djangorestframework
